Rename 'type' to 'component_type' for validation
This is an implicit field of a bolt or spout indicating which it is. I
plan to re-use 'type' to refer to a component being a python component or
otherwise.

The imminent use is to introduce a 'kafka' spout type, but in the future,
it could be used to indicate Java or other kinds of components.