Specify kafka dependency
Between the inclusion of storm-kafka in Storm itself and the release of
0.9.2-incubating, the dependency on kafka was reduced to 'provided'. Thus,
when we upgraded the Storm dependency from 0.9.2-incubating-SNAPSHOT to
0.9.2-incubating, the kafka and transitive scala dependencies were no
longer included in the Pyleus base jar.

This change adds an explicit dependency on kafka_2.9.2 with exclusions for
zookeeper and log4j. This is perhaps not ideal---should we also have
multiple distributions for different versions of scala?---but it's good
enough for now.